requirements.txt
faker>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
neo4j>=5.14.0
torch>=2.1.0
torch_geometric>=2.4.0
langchain>=0.1.0
langgraph>=0.0.10
langchain-openai>=0.0.2
streamlit>=1.30.0
plotly>=5.18.0
networkx>=3.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
pypdf>=3.17.0
pytest>=7.4.0
